Stop relying on an `Arc` dropping for shutdown tasks.
`MT` instances always live within an `Arc`, which we clone and pass
around (e.g. to worker threads, and guards, etc.). That means that when
the main interpreter thread finishes and drops its clone of the `Arc`,
the reference count might not go to zero, and "shutdown" tasks (e.g.
printing stats; checking for worker thread errors) aren't performed.

In retrospect, tying these tasks to `Drop` wasn't a great idea, because
an interpreter author has no idea when an `MT` instance will actually
drop.

This commit introduces an explicit `shutdown` method which interpreter
authors can (should!) call to initiate shutdown tasks. I debated whether
this should also be called from `Drop`, but decided against that,
because new users might see that shutdown tasks are sometimes called
automatically and infer that they're always called automatically. It's
better for us to consistently put the burden on the interpreter author
(whether in C or Rust) to do so explicitly.