Make the JIT IR parser create -ve `Direct` offsets.
Right now, as far as we know, `yksmp::Loaction::Direct` can only have
negative offsets. We really need to check if that is always the case (I
can create guesses for why positive offsets are possible, but I haven't
yet seen a case which does so).

In the meantime, ensure that the "dummp" `Direct` information the JIT IR
parser creates is always negative to match what we see in "real" code.