Kill the notion of running the C suite at different AOT opt levels.
For now we only support -O0 anyway.

In discussions earlier today we agreed that any given tests should be
tailored to one specific opt level.

If/when the time comes to allow other opt levels, we should add a
self-contained way for a test to specify the opt level it requires.