Start to put guard failure transitions in `transition_guard_failure`.
Previously `deopt.rs` did some of this (calling `inc_failed`) and
`mt.rs` did the rest. It took me a while to get my head around that,
because it's subtly different to `transition_control_point`, which is
more self-contained.

This commit moves more of this into `mt.rs`. This then highlights two
inconsistencies:

  1. `mt.rs` controls the `Location` state-machine but here the `Guard`
     itself controls its own state-machine.

  2. Having `Location` in `/location.rs` but `Guard` in
     `/compile/mod.rs` doesn't feel quite right.