Fix memory corruption in deopt.
When copying the reconstructed frame onto the "live" stack, the src
pointer was in RSI, but the call to memcpy(3) can trash it. We then hand
the potentially trashed pointer to free(3). Chaos ensues.

This change fixes this. We stash the src pointer in an unused
callee-save register and restore it before the call to free.

This fixes the Lua AWFY suite for us. I've also the yk and yklua tests a
few times too. Seems OK.