Turn the trace optimiser on by default.
I didn't previously suggest doing this, because there seemed to be a bug
in trace optimisation. That wasn't the case: the bug was elsewhere (ISTR
it was the recent stackmaps issue), and I could (eventually) trigger it
with and without the trace optimiser.

So now seems a good a time as any to switch trace optimisation on by
default. We don't do very many optimisations yet, but by making it the
default we're ensuring that we test the optimiser as it gradually
expands into something more effective.