Use labels to keep track of live variables at loop edges.
When looping back to the top of a trace we need make sure that live
values are put back into the place where the beginning of the trace
expects them. This is because the register allocator may have moved them
somewhere else, e.g. via spilling.

The way we did this until now was by remembering the live values and
their locations at the beginning of the trace, and then restoring them
before jumping at the end of the trace. This isn't entirely correct, but
works as long as all live values live on the stack, because changes to a
value are instantly stored back on the stack.

Once we start using mem2reg, which moves live variables from the stack
into registers, this will no longer work. Changes to variables inside
the loop will no longer be stored on the stack, but instead use PHI
nodes. For example, a loop iterator may look like this:

```
%0: i32: load_ti Register(3, 8, 0, 0)
tloop_start:
%2 = phi [%0, bb0], [%3, bb1]
%3 = add %2, 1i32
jmp tloop_start
```

The increment to the iterator `%0` is stored in a new variable `%3`, and
then the the phi node `%2` picks up the new value on each iteration. If
we only remember the live variables at the beginning of the trace, and
restore them at the end, we would be restoring `%0`, which restores the
old value and causes an inifinte loop.

To fix this we need to track the live variables at the beginning **and**
the end of the trace using labels. Then before we do the backedge jump
we compare them. For any variable where the two locations don't match,
we need to copy the value from the end location to the start location.
For example, if we look at the IR for `simple.c` we now get this:

```
%0: ptr = load_ti Register(3)
%1: ptr = load_ti Direct(6, -40, 8)
%2: ptr = load_ti Direct(6, -32, 8)
%3: ptr = load_ti Register(5)
%4: ptr = load_ti Register(15)
tloop_start: [%0, %1, %2, %3, %4]
...
tloop_jump: [%0, %1, %2, %3, %17]
```

We can already see that the due to SSA the value stored in `%4` is now
stored in `%17`. If we check their locations we might get something
like:

```
%4 => Register(15)
%17 => Register(12)
```

This means we now need to copy the value from R12 into R15, which
ensures that if `%17`'s value is different, this change is propagated to
the beginning of the loop. (Note that since we don't have enabled
mem2reg yet, these changes don't make any difference now, since both
`%4` and `%17` are stack pointers and are already equivalent).