Make `inst_all` a private function.
`inst_all` is a dangerous -- or, perhaps more accurately, an easily
misunderstood -- function, because it allows one to forget about
`ProxyInst`s. Since we've recently introduced `inst_deproxy`, which is
nearly always what we should be using, there are no external uses for
`inst_all`, so we can make it private.