Add the ability for tests to specify an AOT opt level.
This doesn't fit well with the current test harness, so it's a bit of a
hack.

The test file needs a line at the top like:
// ## yk-config-env: YKB_AOT_OPTLEVEL=1

The test harness scrapes this out of the test file and applies the extra
environment to the yk-config invocation.

We don't yet allow a custom pipeline to be specified by a test. Let's
see how far we get with this first.