Fix deproxying.
This was subtly broken (my fault!) before in two ways:
`map_packed_operand_local`s is necessary for (our current) dead code
elimination, but we also (incorrectly) used it elsewhere; and we didn't
use the updated `iidx`s after calling `inst_deproxy`.

This commit fixes both aspects. However, I'm not sure this is the best
long-term approach, and that perhaps the better thing to do is to
continually update (i.e. deproxy) operands as we go along. That can wait
for a later day once I've chewed this over sufficiently, however.