Fix stack overflow when asking for an AOT operand's type.
When adding more passes to the AOT pipeline we found that the mere act
of printing the AOT module can cause stack overflows.

If you have IR like:

```
bb0:
    %1 = ...
bb1:
    %2 = phi bb1 -> %3, bb0 -> %1
    %3 = %2 + 1
    goto bb1
```

Asking the type of %3 causes infinite recursion.

This is because an operation often asks its operands for their type to
determine their own type.

Here, %3 asks %2, who asks %3, who asks %2, who asks... forever. Each
"ask" is a function call, so this blows the stack.

This change breaks the cycle by putting a PHI's type into the
instruction explicitly, so it doesn't have to search further afield to
know it's type.

We haven't seen this before because usually the first operand to a PHI
is from earlier (not later) in the module, so there is no cycle.