O1 fixes.
This change is a collection of fixes required for us to enable the -O1
AOT post-link pipeline in Yk. This is slightly different from just
passing -O1, but it's a big step in the right direction.

Probably the most significant change is the transition from
llvm.experimental.stackamp to llvm.experimental.patchpoint for the
control point. This is needed because we were tracking live values at a
slightly incorrect point in time. We needed the register state
immediately before the call to _ykrt_control_point, but *including* any
register shuffling/spills required due to the call ABI.

The lua_array_len.c test was removed because we now run enough
optimisations for LLVM to statically detect UB in it. Hardly surprising
since this was derived using cvise. We can't handle poison values in the
JIT and even if we could, I doubt the test is a good one any more.

Co-authored-by: Lukas Diekmann <lukas.diekmann@gmail.com>