Reimplement side-tracing.
This commit fixes side-tracing which we disabled when we replaced the
traceinputs struct with passing live variables directly via
registers/stack offsets.

In addition this also implements side-trace looping, i.e. when a
side-trace reaches the end without any guards failing we no longer
deoptimise back to the interpreter but instead loop back to the root
trace.