Get rid of `ykconfig -ao`.
Not so long ago we added `yk-config -ao <level>` to allow us to run our
pre-canned franken-O1 pipeline for the entire test suite. This was only
ever designed as an interim hack, and we've wanted Yk to simply work
with standard optimisation flags like -O1, -O2 etc.

This change is a step in that direction.

We remove `-ao` and our non-standard defaults and add the ability for
individual tests to specify extra compiler flags. This gives us the
ability to specify optimisation parameters on a per-test basis.

I then ported as many tests as were practical to use -O1. The remaining
tests (that were either too involved for me to convert right now, or
actually rely on an IR shape generated by our old default pipeline) I've
left using our franken-O1 by passing:

-O0 -Xclang -disable-O0-optnone -Xlinker \
    --lto-newpm-passes=instcombine<max-iterations=1;no-use-loop-info;no-verify-fixpoint>

We do have to disable yklua in CI for now, as yklua still uses
`yk-config -ao` and we have to break a circular dependency.