Rewrite the optimiser to be a one pass optimiser.
[Well, OK the first lie is in the first line. Because we have to do
backwards code generation, we actually have to do a second pass. But
that's a boring detail, and something that predates this commit.]

This commit moves the optimiser into a model that's closer to RPython:
rather than multiple forward stages each doing one optimisation, we have
a single forward stage that does all optimisations in one go. This is
quicker (fairly obviously!) but also simpler, because it means that we
can make strong assumptions at the point of the wavefront: no-one can
change our knowledge of earlier parts of the trace.

The crucial difference between this commit and what came before is that
as well as mutating the IR as we go along (e.g. constant folding) we
maintain an analysis of what values an instruction can produce, which we
gradually refine. Crucially, mutation can only happen at the current
instruction, but analysis refinement can happen on previous
instructions.

Consider a trace along these lines:

```
%0: i8 = load_ti ... ; at this point we know nothing about %0
%1: i8 = add %0, 1i8 ; we still know nothing about %0
%2: i1 = eq %0, 0i8  ; we know that henceforth %0 must be 0i8
guard true %2
%4: I8 = mul %0, %0  ; we know this must equal the constant 0i8
```

Notice that it's only at the point of the guard that our analysis allows
us to start using the knowledge "%0 = 0i8" for optimisations.

The code is currently woefully incomplete (but if I fill in anything
else, I trip up on the "removing guards tends to kill things" bug that I
believe is fixed but not yet merged), I haven't yet bothered porting
across one old optimisation (`mul_chain`), and I haven't really got a
good API for sharing analysis and mutations (which is currently done
inconsistently).

All that said, this nudges the optimiser forward just about enough that
I think it's worth considering for merging now, and then it can be
improved in-tree.