Optimise codegen for 64-bit addition with a constant that fits in 32-bits.
I started out trying to cover too many cases any getting confused, so
this optimises only one very specific case:

    %2: i64 = add %1, <const>

Where the numeric value of <const> is expressible in 32 bits.

For this we use the instruction encoding:
  add r64, imm32

This builds on a sketch @ltratt started and handed to me to look at. His
work also contains some semi-related promote improvements also.

I had hoped to make a series of self-contained commits, but the work wen
through so much flux that it didn't work out, and would only force the
reviewer to review stale code.

Here's some miscellaneous notes on the changes:

 - Fix a bugs in `Truncate` and `SignExtend`.
 - Unit test and document `Truncate` and `SignExtend` too.
 - Fix incorrect bit size computation (byte_size * 8 isn't what we want).
 - Unit test the optimisation.
 - Add C test to check runtime semantics of the optimisation.
 - size -> byte_size, bits -> bit_size (for improved comprehension)

All lua tests and benchmarks work.

bigloop.lua uses this optimisation in the trace it generates.