Emit add instructions with immediate operands where possible.
Not all tests pass yet. Read on.

An explanation of why I think this is correct:

Suppose we have the JIT IR instruction:

 %2: i32 = add %1, -1i32

Here, -1 is constant, and therefore we might be able to avoid loading it
into a register.

The X64 ISA has no `add r/m64, imm64`, but it does have:
 - add r/m32, i8
 - add r/m32, i32

We can use the former instruction if our constant fits in a i8.
We can use the latter instruction if our constant fits in a i32.

As it happens, -1 is expressible for all i<n> by setting all <n> bits
(i.e. all ones). So -1u8 is simply 0xff and -1u32 is 0xffffffff. This
means we could in fact emit either of the above instructions.

Let's use the first instruction encoding above. In the JIT IR, constant
integers are stored as a u64, so we start by throwing away the undefined
top bits with `let const_i32 = const_val as i32`. This gives us -1i32.

Then we check if we can narrow the type to i8 whilst retaining the same
numeric value with `let const_i8 = i8::try_from(const_i32)`. As
explained above this succeeds and we get -1i8 (0xff).

Supposing the result is to be stored in rax, we then emit the following
instruction:

  add rax, BYTE const_i8

This approach only works for "normally-sized" integers like i8, i16,
i32 and i64 for now. To use immediate operands for odd integer types
like i3 we'd have to emit extra instructions to sign-extend the constant
value up to the next usable bit-width with the appropriate instruction
encoding. That's one for another day.

bigloop.lua runs under yklua and generates the correct value, but here's
the mystery: this change consistently slows down yklua jitting
bigloop.lua by about 8% (that's a release build of yklua with serialised
compilation, master branch vs. this branch).

At first I thought it was due to register stalls (when you write to 8-
or 16-bit registers), but if you look at my diff, those cases are
todo!() and thus are not covered by bigloop.lua and therefore not the
problem.

The trace contains 4 JIT IR add instructions:
 - 2 adds involve a constant that we "optimise". The constant values are -1 and -127.
 - 2 adds do not involve constants. Sadly this must include the blatent
   addition evident from the benchmark source code:

   `sum = sum + 1`

I guess we need to teach Lua to promote language level constants.
@ltratt has talked about this problem before.

So I don't know why this slows us down. Do you see anything fishy? Am I
using the register allocator correctly?