Avoid use of 8 and 16-bit registers in x64 codegen.
The codegen routines now try to avoid the use of operations which set 8
or 16-bit sub-registers. This is for two reasons: a) to simplify the
code, but also b) to avoid partial register stalls in the CPU pipeline.
Generally speaking we prefer to use 64-bit operations, but sometimes we
special-case 32-bit operations since they are so common.

If an object occupies only part of a 64-bit register, then there are no
guarantees about the unused higher-order bits. Codegen routines must be
mindful of this and, depending upon the operation, may be required to
(for example) "mask off" or "sign extend into" the undefined bits in
order to compute the correct result.