Adjust the root jump target for side-traces.
When side-traces jumped back to the root trace, they previously always jumped to
the entry address of the root trace. Since we want to implement loop peeling
soon, we need to be able to jump to an offset inside the root trace. For now
this is simply the place after the trace's prologue. But later we want to change
this to the place after the preamble and before the peeled loop.

This change required some additional changes to the way we calculate and reset
the base pointer when jumping to the root trace.