If a guard's live variable refers to an operand, make it a constant.
It is common for guards' live variables to refer to the operands e.g.:

```
%18: i1 = sgt %17, 0i32
guard true, %18, [0:%0_6: %0, 0:%0_7: %1, 0:%0_8: %2, 0:%0_9: %3, 0:%9_1: %18]
```

where `0:%9_1: %18` is the live variable referring to the operand.

This might seem harmless, albeit a bit pointless. However, it's actually
a pain in the rear end, because we may want not to assign a
register/stack location to `%18`, but we have to because it's referred
to by an AOT live variable. In particular, for an upcoming change where
we want to optimise machine code across ICmp/Guard instructions, this
dependency is problematic.

This commit recognises that there's no need for this: if the guard
fails, we know the value of the operand (for `guard true` it must be 0;
for `guard false` it must be 1). This allows us to break the link: we
change the guard to:

```
%18: i1 = sgt %17, 0i32
guard true, %18, [0:%0_6: %0, 0:%0_7: %1, 0:%0_8: %2, 0:%0_9: %3, 0:%9_1: 0i1]
```

I've added a new test (`guard_consting.c`) which enforces this property.