Optimise `icmp` immediately followed by `guard`.
Previously if we had a comparison followed by a guard, we generated code
such as:

```
; %60: i1 = eq %59, 3i8
mov rdx, 0x03
movzx rbx, bl
movzx rdx, dl
cmp rbx, rdx
setz bl
; guard true, %60, [....]
cmp bl, 0x01
jnz 0x00000000000004AD
```

In other words, we generate an intermediate value (clobbering a
register!) and immediately consume it. This commit changes the above
code to:

```
; %60: i1 = eq %59, 3i8
movzx rdx, dl
cmp rdx, 0x03
; guard true, %60, [...]
jnz 0x000000000000042E
```

The `movzx` is uninteresting from this commit's perspective: the crucial
thing is that the compare/guard are now just two instructions (`cmp` &
`jnz`). This gives a 10% boost to big_loop.lua.

In order to generate this code, the x64 backend explicitly recognises
that it can "merge" code across the two opcodes: note that not all
backends might be able to do this, so it feels right (at least at the
moment!) to do it in the backend and not introduce a new JIT IR
instruction.

Whilst I'm here, factor out and generalise the "try to use
in-instruction immediates if a constant fits", which slightly improves
our codegen elsewhere (hence the tweak to the `cg_const` test).

Note that there is more optimisation potential for this sort of thing.
For example it's not uncommon that the icmp and guard are separated
e.g.:

```
%10: i1 = eq ...
%11: i8 = add ...
guard true %10, ...
```

This commit falls back on the old, verbose and slow, code path in such
cases. We can obviously do better, though which of the possible design
options for doing so is best I'm not yet entirely sure.