Remove now-unnecessary hack that keeps variables alive.
Before we had loop jumping, we didn't really know how long to keep
variables alive for: we thus added a temporary hack (ha!) to keep alive
all trace inputs until the end of the trace. After we added loop
jumping, this became unnecessary -- but, of course, we all (including
me!) had forgotten the temporary hack ever existed!

This commit removes the hack, which I accidentally stumbled across when
I found a test I couldn't express because of it.

Removing this has a much more positive effect than I expected:

  1. We remove quite a bit of pressure from the register allocator
     (always a good thing)
  2. We don't have to restore anywhere near as many variables at the end
     of a trace (some of which we had to pop from the stack, which is
     always slow).
  3. In tests, we no longer add as many mystery "do something with the
     stack" instructions, making tests easier to understand.

Most of this commit just adjusts the test suite because of (3). In a
couple of cases, the tests now become a bit more verbose, because the
old tests accidentally relied on the register allocator keeping
registers alive too long and thus making registers accidentally "unique"
because it couldn't reuse them.