Don't require updating the running trace during side-tracing.
In order to find the correct information during a guard failure we had
to switch the running trace whenever we jumped from a root trace into a
side-trace and back. Not only does this slow us down, it occasionally
lead to bugs when trying to find the root trace.

This commit no longer requires updating the running trace and instead
passes down a list of all the guards that have failed leading up to the
current guard failure. This list can then be used to find the correct
running trace and guard failure information by traversing down from the
root trace.