Fix a side-tracing + hardware tracing bug.
When we are constructing a side trace we will have first deopted back to
the conditional branch that initiated side-tracing. Because the
conditional branch is re-executed, this means that when we are doing
hardware tracing, we will likely see a spurious blocks at the beginning
of the trace that are not desirable in the side-trace. We therefore skip
them.

Fixes crashes like this, where we get a local_map miss for local
variables relating to the conditional branch:

thread '<unnamed>' panicked at ykrt/src/compile/jitc_yk/trace_builder.rs:404:69:
no entry found for key

Co-authored-by: Lukas Diekmann <lukas.diekmann@gmail.com>