Make clear that parameter indexes are `InstIdx`s.
Previously we used a `u32` (why such a big type? dunno!), requiring
`ParamInst` to be `packed`. By definition a parameter is stored in an
SSA variable which, in our context, means it's an `InstIdx`.

This commit makes this explicit:

  1. Parameters are now indexed by `InstIdx`.
  2. The vague `locidx` is now `paramidx`.
  3. The unnecessary largesse of the `params()` function returning a
     slice is hidden by a `param()` function which returns a single
     parameter.

This change did spot a bug in a test, so it is useful.