Update the FP register allocator.
In essence, this commit ports over the GP register allocator to the FP
register allocator. It would be nice to factor out the common code at
some point, but not today.

This fixes an allocator crash in `fmul` (which was probably indicative
of a class of similar problems in FP register allocation). Bug found
by Edd Barrett.

Co-authored-by: Edd Barrett <vext01@gmail.com>