Use a pinned version of yklua in `.buildbot.sh`.
Previously we have got ourselves into trouble when yklua has changed,
and it causes later yk builds to break. There's no guaranteed way around
this problem, but we can avoid most of the pain by forcing ourselves to
use a specific version of yklua in tests.

This commit does that for `.buildbot.sh` (and, while I'm here, I've done
a two-stage-clone-checkout to minimise network traffic). We could
perhaps use a submodule for this, but since we don't use yklua anywhere
else in the test suite, that feels a bit intrusive.