Recast CSE as iterating over an instruction-specific linked list.
In essence this commit ports a data structure -- which also leads us to a
less fiddly API -- from LuaJIT. As we're optimising a trace we create a
linked list of all instructions of type T. In common subexpression
elimination we can then look backwards only at instructions of type T.
This is faster -- we move from O(n) best case to O(1) best case, though
the worst case of O(n) necessarily remains -- but also means we can hide
quite a bit of API nastiness.

The data structure itself is quite cunning (and I take no credit for
this). We have `heads`, which contains one entry per `Inst` discriminant
(i.e. this is a fixed-size array) and `predecessors` with one entry per
instruction in a trace: `heads` tells you where to start in the linked
list, and `predecessors` actually records the linked list itself. Thus
creating this data structure requires a single memory allocation (for
`predecessors`).