Move the CSEing of guards into the `cse` function.
We (have to) use a slightly relaxed form of CSE for guards (we don't
reference their `GuardInfo`s, which by definition will be different).
Previously we handled this in the main part of the optimiser, but this
really is most obviously thought of as a form of CSE, so moving it here
makes it clearer as to what's really going on.