Propagate information about the register an operand should end up in.
In essence this is a poor man's version of reverse code generation:
during the reverse analysis phase, we propagate information about the
desired register for an operand's output backwards starting from
`tloop_jump`. In the medium term we should also propagate information
about spill locations, but this commit only propagates information about
registers.

This is already a win of sorts -- 2-3% on big_loop -- but our
frequent use of the `InputOut` constraint means that the register
allocator often ends up `mov`ing unnecessarily, slightly defeating the
purpose of our hints. I have known for a while that this was suboptimal,
but couldn't find a nice API for it, so haven't addressed it before.
With luck, changes such as that will make the register hints more
effective.