Add `OutputCanBeSameAsInput` constraint.
This is useful for 3-operand instructions, where one can implicitly
avoid clobbering an input variable if that is going to be used later.
For example imagine we have an input variable in `r10` and it will be
used in the next instruction. Before we would have had to do this for a
`PtrAdd`:

```
mov r11, r10 # r11 unused so we can move r10 to avoid clobbering it
add r10, 0x20
```

After this PR it makes sense to convert `add` to `lea` and do the
declobbering differently:

```
lea r11, [r10+0x20]
```

Note this leaves the value in `r10` unclobbered.

We had a hack for `cg_load` which did this in an unsatisfying way, but
this commit moves things into the register allocator where they belong.
The new `OutputCanBeSameAsInput` constraint allows the register
allocator to use the same register as an input operand, or not, as it
sees fit. `cg_ptradd` shows this clearly:

```rust
let ptr_op = inst.ptr(self.m);
let [in_reg, out_reg] = self.ra.assign_gp_regs(
    &mut self.asm,
    iidx,
    [
        RegConstraint::Input(ptr_op.clone()),
        RegConstraint::OutputCanBeSameAsInput(ptr_op),
    ],
);

dynasm!(self.asm ; lea Rq(out_reg.code()), [Rq(in_reg.code()) + inst.off()]);
```

Depending on whether `ptr_op` is used later in the trace, `in_reg` and
`out_reg` might reference the same, or different, registers.

This squeezes another 2-3 percentage points out of big_loop.